+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Only accretive/fixative changes will be made from now on.
 
 * 1.2.next in progress
+  * Address #193 by expanding the argument specs for `get-datasource` and `get-connection`.
   * Update log4j2 test dependency.
   * Update `build-clj` to v0.6.5.